Weird performance cliff using a `Slice`
The sequence passed to `String(bytes:encoding:)` has to be the correct
length -- the `init` doesn't just stop at `NUL`. If we slightly modify
the `rawMorePointer` script so that it returns a `Slice` of the working
buffer to match `String` requirement, the script is as slow or slower
than the`Character`-based version! (An order of magnitude slower than
the original `rawMorePointer`.)