Skip non-digits between braces
Stop parsing if there are characters that aren't hexadecimal digits.
This would be an error condition in a real parser; for these tests we
just continue as if the bad sequence were normal text, appending it to
the output.

Switch the C implementation to use stdlib functions for digit parsing. I
attempted to use `strtol` for the `rawMorePointer` version as well, but
was defeated by the need to turn an `UnsafePointer<UInt8>` into an
`UnsafeMutablePointer<UnsafeMutablePointer<Int8>?>?` for `strtol`'s `end`
argument. So we still do the digit parsing in Swift there.

The extra checking has minimal effect on speed, though of course
there are no invalid sequences in the test input.