Harden processing against bad input
The C implementation in particular skipped some error checking that Swift
code is doing implicitly, which could produce an unfair speed comparison.
Used American Fuzzy Lop to help identify some of these faults.

On finding a hex sequence whose value is too high to be a valid
codepoint, the whole escape sequence is treated as normal text, instead of
the program aborting. This also applies to rawMorePointer.swift.

Check input for validity as a UTF-8 C string before attempting to process
it, which Swift `String` already does. Not surprisingly, this has a small
but noticeable impact on speed for the standard test input size.